Fix : Overflow issues in array_patterns

+++ c/array-patterns/array10_pattern.c
@@ -29,11 +29,11 @@ void __VERIFIER_assert(int cond) { if(!(cond)) { ERROR: __VERIFIER_error(); } }
 extern int __VERIFIER_nondet_int() ;
 extern short __VERIFIER_nondet_short() ;
 
-short ARR_SIZE ;
+signed long long ARR_SIZE ;
 
 int main()
 {
-	ARR_SIZE = __VERIFIER_nondet_short() ;
+	ARR_SIZE = (signed long long)__VERIFIER_nondet_short() ;
 	__VERIFIER_assume(ARR_SIZE > 0) ;
 
 	int array1[ARR_SIZE] ;
@@ -41,7 +41,7 @@ int main()
 	int count = 0, num = -1 ;
         signed long long sum = 0 ;
 	int temp ;
-	short index ;
+	signed long long index ;
 
 	for(count=0;count<ARR_SIZE;count++)
 	{
@@ -52,7 +52,7 @@ int main()
 	while(1)
         {
 		
-		index = __VERIFIER_nondet_short() ;
+		index = (signed long long)__VERIFIER_nondet_short() ;
 		__VERIFIER_assume(index>=0 && index < ARR_SIZE) ;
 
 		array1[index] = num * (num * (num * (num *index))) ;
